booking-service/requirements.txt
mysql-connector-python==9.1.0
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1